test(metadata): consolidate test files and collapse iterative clusters
Merge the three previous test files into the canonical metadata_test.go
and table-ify the per-role clusters where the bodies share a template.

File-level consolidation:
* metadata_table_test.go (table-driven I/O coverage) + the existing
  metadata_api_test.go (API-level tests, TestMain) merge into
  metadata_test.go. No logic changes from this part -- pure textual
  concatenation with merged imports.

In-file consolidation:
* TestIsExpiredRoot/Snapshot/Timestamp/Targets (4) -> TestIsExpiredTable
  (1 test, 4 cases).
* TestRootReadWriteReadCompare/Snapshot.../Targets.../Timestamp... (4)
  -> TestRoundtripFileTable (1 test, 4 cases) using a per-role
  closure that reads a fixture, writes it, re-reads, and compares.

marshal_test.go, logger_test.go, metadata_bench_test.go, and
metadata_fuzz_test.go remain as-is: they cover distinct concerns
(JSON marshalling edge cases, logger plumbing, benchmarks, fuzz
corpora) and grouping them into metadata_test.go would obscure
their purpose without adding coverage. The other iterative-looking
clusters (TestDefaultValues*, TestVerifyDelegateDuplicate*,
TestVerifyLengthHashes*, TestUnrecognizedField*) are deliberately
left iterative because the per-case bodies differ enough that a
table would obscure rather than clarify.

Net: 66 -> 60 Test* functions; coverage holds at 82.0% of metadata
statements.
